systemstats: Consider discharge cap in remaining time calculation
The battery is typically configured not to discharge below 10%,
which skewes the remaining time calculation since there's an
additional half-a-kilowatt or so remaining that will not acutally
be used.

Rather than introduce a fully-fledged wrapper around the
getDisChargeConfigInfo (which also includes configurable time
settings), just call it directly for this one particular use case.